485776: Fix file watch issues
We have been seeing file watch issues. This work setsup Webpack to run
preDev, and to set a watcher on nodemon to listen for non-webpack files.

This feels more reliable than the previous implementation of watching
the public folder for changes from webpack. It also means we can drop
the webpack watcher poll option which wasn't being used and decrease the
aggregateTimeout to 0. Whic h should speed some things up a little.

I've been sat with this all day and its behaved itself so getting it in
to the codebase