CORE-416: enable hint order/position within suggestion
Enable hint ordering within a suggestion. Allowing devs to provide
different order/position to the hint in suggestions.

fEx for the hint in the suggestion for test runs it appears as the first
item in the suggestion. In the hint in the image version during the
deployment flow it appears last.

This is controlled via order in a css class thats now passed through to
the suggestions element